Fix code of pair parser combinator to pass the tests
There is a bug in pair parser combinator that causes introduced tests to fail.
The problem is that when the first parser successfully parses the input
and the second doesn't, the combinator returns the error representing
the input of the second parser not the input of the whole pair
combinator.